% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_scan.R
\name{augment.kaplan_scan}
\alias{augment.kaplan_scan}
\title{Augment data with information from a kaplan_scan object}
\usage{
\method{augment}{kaplan_scan}(x)
}
\arguments{
\item{x}{A kaplan_scan object}

\item{...}{Extra arguments, not used}
}
\value{
A tibble with 4 columns: time, event, expression and expression_level.
}
\description{
Augment data with information from a kaplan_scan object
}
