% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_scan.R
\name{best_threshold}
\alias{best_threshold}
\title{Extract the threshold expression value able to achieve the biggest survival curve separation upon dichotomization.}
\usage{
best_threshold(time, event, expression)
}
\arguments{
\item{time, event, expression}{Numerical vectors containing the follow up time, status indicator and gene expression.}
}
\description{
Uses the KaplanScan algorithm to determine the threshold.
}
