% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_scan.R
\name{kaplan_scan}
\alias{kaplan_scan}
\title{Stratify participants using the KaplanScan algorithm.}
\usage{
kaplan_scan(data, time, event, expression, p.adjust.method = "bonferroni")
}
\arguments{
\item{data}{A data frame.}

\item{time, event, expression}{Names of the columns containing the follow up time, status indicator and gene expression.}

\item{p.adjust.method}{p value correction method (see \link[stats]{p.adjust}).}
}
\value{
\code{kaplan_scan} returns an object of \link[base]{class} "kaplan_scan"
}
\description{
Use the KaplanScan algorithm from the \href{http://r2.amc.nl/}{R2 Genomics Analysis and Visualisation Platform}
}
\details{
Implemented as described in the \href{http://hgserver1.amc.nl/r2/help/manual/R2-Manual.html#section_5.2.5}{R2 manual}: \emph{To this end, we start by ordering the samples within a dataset from the lowest expressor to the highest expressor for a given gene. In the next step we start at the lower end of expression and determine the logrank test for the first 8 samples against all the higher expressed samples. We then increase the group by the addition of the 9th sample to the first group and repeat this procedure until the second group contains 8 samples. In the final step the best logrank test is chosen to represent the optimal cut-off point for the current gene. As we are performing a whole series of tests to find the most optimal cut-off, all the p-values of the logrank tests are corrected for multiple testing by bonferoni.}
The procedure describesd above can be performed iteratively to all genes on the array, which would thereby identify the most significant logrank tests. This is exactly what the kaplan scan for a group of genes does.
}
