% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_scan.R
\name{ks_coxph}
\alias{ks_coxph}
\title{Fit Proportional Hazards Regression Model for the survival data}
\usage{
ks_coxph(time, event, expression)
}
\arguments{
\item{time}{Follow up time}

\item{event}{The status indicator, normally 0=alive, 1=dead.}

\item{expression}{The expression stratus (normally 'low' vs 'high')}
}
\value{
an object of class coxph representing the fit. See coxph.object for details.
}
\description{
Fit Proportional Hazards Regression Model for the survival data
}
\details{
Define a standalone function to pretty print the model in the kaplan_scan object
}
